+++ comics/comics/fanboys.py
@@ -4,18 +4,19 @@ from comics.meta.base import MetaBase
 class Meta(MetaBase):
     name = 'F@NB0Y$'
     language = 'en'
-    url = 'http://fanboys-online.com/'
+    url = 'http://www.fanboys-online.com/'
     start_date = '2006-04-19'
     rights = 'Scott Dewitt'
 
 class Crawler(CrawlerBase):
     history_capable_days = 180
+    schedule = 'Mo,Fr'
     time_zone = -5
 
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://fanboys-online.com/?feed=rss2')
+        feed = self.parse_feed('http://www.fanboys-online.com/?feed=rss2')
         for entry in feed.for_date(pub_date):
-            if 'Comic' not in entry.tags:
+            if 'Comics' not in entry.tags:
                 continue
             url = entry.summary.src('img.comicthumbnail')
             if not url: