+++ docs/changes.rst
@@ -15,6 +15,7 @@ v1.1.5 (in development)
 - Update: ``calamitesofnature`` has been discontinued.
 - Update: ``duelinganalogs`` was broken due to feed changes.
 - Update: ``fagprat`` has a new schedule.
+- Update: ``fanboys`` was broken due to feed changes.
 
 
 v1.1.4 (2012-04-07)